+++ js/panorama.js
@@ -52,6 +52,7 @@ OCA.Analytics = Object.assign({}, OCA.Analytics, {
     currentReportData: {},
     isPanorama: true,
     // flexible mapping depending on type required by the used chart library
+    // Add in all js files!
     chartTypeMapping: {
         'datetime': 'line',
         'column': 'bar',
@@ -59,7 +60,8 @@ OCA.Analytics = Object.assign({}, OCA.Analytics, {
         'columnSt100': 'bar', // map stacked type also to base type; needed in filter
         'area': 'line',
         'line': 'line',
-        'doughnut': 'doughnut'
+        'doughnut': 'doughnut',
+        'funnel': 'funnel'
     },
     headers: function () {
         let headers = new Headers();